Refactor blobfinder, splitting up in several files
