+++ docs/source/changelog/misc/blobfinder.rst
@@ -1,0 +1,4 @@
+[Misc] Restructure blobfinder
+=============================
+
+* Split up the bobfinder code to reduce file size (:pr:`468`)
