+++ docs/source/changelog/misc/testperformance.rst
@@ -1,4 +1,4 @@
 [Misc] Test performance
------------------------
+=======================
 
 * Performance improvements for unit tests (:issue:`447`, :pr:`448`)
