+++ src/libertem/udf/blobfinder/__init__.py
@@ -1,0 +1,21 @@
+from .patterns import (
+    MatchPattern, Circular, RadialGradient, BackgroundSubtraction, UserTemplate,
+    RadialGradientBackgroundSubtraction
+)
+from .refinement import (
+    RefinementMixin, FastmatchMixin, AffineMixin, run_refine
+)
+from .correlation import (
+    get_peaks, CorrelationUDF, FastCorrelationUDF, FullFrameCorrelationUDF,
+    SparseCorrelationUDF, run_fastcorrelation, run_blobfinder
+)
+from .utils import feature_vector, visualize_frame, paint_markers
+
+__all__ = [
+    'MatchPattern', 'Circular', 'RadialGradient', 'BackgroundSubtraction', 'UserTemplate',
+    'RadialGradientBackgroundSubtraction',
+    'RefinementMixin', 'FastmatchMixin', 'AffineMixin', 'run_refine',
+    'get_peaks', 'CorrelationUDF', 'FastCorrelationUDF', 'FullFrameCorrelationUDF',
+    'SparseCorrelationUDF', 'run_fastcorrelation', 'run_blobfinder',
+    'feature_vector', 'visualize_frame', 'paint_markers'
+]
